@@ -1,62 +1,59 @@
 from __future__ import annotations
-from dataclasses import dataclass
-from typing import Optional, Tuple
-
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
 from .backbones.dinov3 import DinoV3Backbone
 from .fine_cnn import FineCNN
-from .heads import HybridHeads, FeatureOutputs
+from .heads import Heads, FeatureOutputs
 
 class LocalFeatureNet(nn.Module):
-    """DINOv3 tokens + fine CNN -> detector+descriptor(+offset,+reliability) at fine stride."""
-
+    """
+    DINOv3 tokens (stride=16) + FineCNN (stride=4) -> detector/descriptor/offset/reliability at stride=4.
+    """
     def __init__(
         self,
+        dinov3_name: str,
         patch_size: int = 16,
-        descriptor_dim: int = 128,
-        fine_channels: int = 64,
-        fine_blocks: int = 6,
-        fine_stride: int = 4,
+        descriptor_dim: int = 256,
+        fine_channels: int = 96,
+        fine_blocks: int = 8,
+        freeze_backbone: bool = True,
         use_offset: bool = True,
         use_reliability: bool = True,
-        freeze_backbone: bool = True,
+        dinov3_dtype: str = "bf16",
     ):
         super().__init__()
-        self.backbone = DinoV3Backbone(patch_size=patch_size, freeze=freeze_backbone)
-        self.fine = FineCNN(in_ch=3, channels=fine_channels, num_blocks=fine_blocks, out_stride=fine_stride)
+        self.patch_size = int(patch_size)
 
-        # DINO tokens will be upsampled to match fine feature resolution
-        # Concatenate fine features (C_f) + dino (C_d)
-        # We don't know C_d until DINO is loaded, so we build heads lazily.
-        self.descriptor_dim = descriptor_dim
-        self.use_offset = use_offset
-        self.use_reliability = use_reliability
-        self.fine_stride = fine_stride
+        self.backbone = DinoV3Backbone(
+            name_or_path=dinov3_name,
+            patch_size=self.patch_size,
+            freeze=freeze_backbone,
+            dtype=dinov3_dtype,
+        )
+        self.backbone.load()
+        assert self.backbone.embed_dim is not None and self.backbone.embed_dim > 0
 
-        self._heads: Optional[HybridHeads] = None
+        self.fine = FineCNN(in_ch=3, channels=int(fine_channels), num_blocks=int(fine_blocks))
+        in_ch = int(fine_channels) + int(self.backbone.embed_dim)
 
-    def _build_heads_if_needed(self, dino_c: int, fine_c: int, device: torch.device) -> None:
-        if self._heads is not None:
-            return
-        self._heads = HybridHeads(
-            in_ch=dino_c + fine_c,
-            descriptor_dim=self.descriptor_dim,
-            use_offset=self.use_offset,
-            use_reliability=self.use_reliability,
-        ).to(device)
+        self.heads = Heads(
+            in_ch=in_ch,
+            descriptor_dim=int(descriptor_dim),
+            use_offset=bool(use_offset),
+            use_reliability=bool(use_reliability),
+            max_offset=0.5,
+        )
 
     def forward(self, x: torch.Tensor) -> FeatureOutputs:
-        # x: B,3,H,W
-        fine = self.fine(x)  # B,Cf,Hf,Wf (Hf=H/stride)
-        dino = self.backbone(x).tokens  # B,Cd,H/patch,W/patch
+        fine = self.fine(x)  # (B,Cf,H/4,W/4)
 
-        # upsample dino to fine resolution
-        Hf, Wf = fine.shape[-2:]
-        dino_up = F.interpolate(dino, size=(Hf, Wf), mode="bilinear", align_corners=False)
+        # DINO tokens at stride 16, then upsample to stride 4
+        with torch.autocast("cuda", enabled=False):
+            dino = self.backbone(x).tokens
+        dino_up = F.interpolate(dino, size=fine.shape[-2:], mode="bilinear", align_corners=False)
+        dino_up = dino_up.to(dtype=fine.dtype)
 
-        self._build_heads_if_needed(dino_c=dino_up.shape[1], fine_c=fine.shape[1], device=x.device)
         feat = torch.cat([fine, dino_up], dim=1)
-        return self._heads(feat)  # type: ignore[arg-type]
+        return self.heads(feat)
